requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
Pillow>=9.0.0
omegaconf>=2.3.0
tqdm>=4.65.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
# Optional: tensorboard>=2.13.0
